refactored so create VirtualBackends
