+++ virtualizarr/backends/api.py
@@ -1,26 +1,17 @@
-import os
 import warnings
-from collections.abc import Iterable, Mapping, MutableMapping
+from collections.abc import Iterable, Mapping
 from enum import Enum, auto
-from io import BufferedIOBase
 from pathlib import Path
 from typing import (
     Any,
-    Hashable,
     Optional,
-    cast,
 )
 
-import xarray as xr
 from xarray import Dataset
-from xarray.backends import AbstractDataStore, BackendArray
-from xarray.core.indexes import Index, PandasIndex
-from xarray.core.variable import IndexVariable, Variable
+from xarray.core.indexes import Index
 
 from virtualizarr.manifests import ManifestArray
 from virtualizarr.utils import _FsspecFSFromFilepath
-
-XArrayOpenT = str | os.PathLike[Any] | BufferedIOBase | AbstractDataStore
 
 
 class AutoName(Enum):
@@ -89,12 +80,6 @@ def automatically_determine_filetype(
     return filetype
 
 
-class ManifestBackendArray(ManifestArray, BackendArray):
-    """Using this prevents xarray from wrapping the KerchunkArray in ExplicitIndexingAdapter etc."""
-
-    ...
-
-
 def open_virtual_dataset(
     filepath: str,
     *,
@@ -107,7 +92,7 @@ def open_virtual_dataset(
     indexes: Mapping[str, Index] | None = None,
     virtual_array_class=ManifestArray,
     reader_options: Optional[dict] = None,
-) -> xr.Dataset:
+) -> Dataset:
     """
     Open a file or store as an xarray Dataset wrapping virtualized zarr arrays.
 
@@ -187,26 +172,26 @@ def open_virtual_dataset(
             filepath=filepath, reader_options=reader_options
         )
 
-    # TODO define these through a pluggable entrypoint system instead
+    # TODO define these through a mapping to registered pluggable entrypoints instead
     match filetype.name.lower():
         case "kerchunk":
-            from virtualizarr.readers.kerchunk import open_virtual_dataset
+            from virtualizarr.readers.kerchunk import KerchunkVirtualBackend
 
-            return open_virtual_dataset(filepath, reader_options)
+            return KerchunkVirtualBackend.open_virtual_dataset(filepath, reader_options)
 
         case "zarr_v3":
-            from virtualizarr.readers.zarr_v3 import open_virtual_dataset
+            from virtualizarr.readers.zarr_v3 import ZarrV3VirtualBackend
 
-            return open_virtual_dataset(
+            return ZarrV3VirtualBackend.open_virtual_dataset(
                 storepath=filepath,
                 drop_variables=drop_variables,
                 indexes=indexes,
             )
 
         case "dmrpp":
-            from virtualizarr.readers.dmrpp import open_virtual_dataset
+            from virtualizarr.readers.dmrpp import DMRPPVirtualBackend
 
-            return open_virtual_dataset(
+            return DMRPPVirtualBackend.open_virtual_dataset(
                 filepath,
                 drop_variables=drop_variables,
                 loadable_variables=loadable_variables,
@@ -215,9 +200,9 @@ def open_virtual_dataset(
             )
 
         case "netcdf3":
-            from virtualizarr.readers.netcdf3 import open_virtual_dataset
+            from virtualizarr.readers.netcdf3 import NetCDF3VirtualBackend
 
-            return open_virtual_dataset(
+            return NetCDF3VirtualBackend.open_virtual_dataset(
                 filepath,
                 group=group,
                 drop_variables=drop_variables,
@@ -227,9 +212,9 @@ def open_virtual_dataset(
             )
 
         case "hdf5" | "netcdf4":
-            from virtualizarr.readers.hdf5 import open_virtual_dataset
+            from virtualizarr.readers.hdf5 import HDF5VirtualBackend
 
-            return open_virtual_dataset(
+            return HDF5VirtualBackend.open_virtual_dataset(
                 filepath,
                 group=group,
                 drop_variables=drop_variables,
@@ -244,9 +229,9 @@ def open_virtual_dataset(
             raise NotImplementedError(f"Unsupported file type: {filetype}")
 
         case "tiff":
-            from virtualizarr.readers.tiff import open_virtual_dataset
+            from virtualizarr.readers.tiff import TIFFVirtualBackend
 
-            return open_virtual_dataset(
+            return TIFFVirtualBackend.open_virtual_dataset(
                 filepath,
                 group=group,
                 drop_variables=drop_variables,
@@ -256,9 +241,9 @@ def open_virtual_dataset(
             )
 
         case "fits":
-            from virtualizarr.readers.fits import open_virtual_dataset
+            from virtualizarr.readers.fits import FITSVirtualBackend
 
-            return open_virtual_dataset(
+            return FITSVirtualBackend.open_virtual_dataset(
                 filepath,
                 group=group,
                 drop_variables=drop_variables,
@@ -269,137 +254,3 @@ def open_virtual_dataset(
         case _:
             raise NotImplementedError(f"Unsupported file type: {filetype.name}")
 
-
-# TODO move these out into a backend/utils.py module? Or readers/common.py?
-def open_loadable_vars_and_indexes(
-    filepath: str,
-    loadable_variables,
-    reader_options,
-    drop_variables,
-    indexes,
-    group,
-    decode_times,
-) -> tuple[Mapping[str, Variable], Mapping[str, Index]]:
-    """
-    Open selected variables and indexes using xarray.
-
-    Relies on xr.open_dataset and its auto-detection of filetypes to find the correct installed backend.
-    """
-
-    # TODO get rid of this if?
-    if indexes is None or len(loadable_variables) > 0:
-        # TODO we are reading a bunch of stuff we know we won't need here, e.g. all of the data variables...
-        # TODO it would also be nice if we could somehow consolidate this with the reading of the kerchunk references
-        # TODO really we probably want a dedicated xarray backend that iterates over all variables only once
-        fpath = _FsspecFSFromFilepath(
-            filepath=filepath, reader_options=reader_options
-        ).open_file()
-
-        # fpath can be `Any` thanks to fsspec.filesystem(...).open() returning Any.
-        # We'll (hopefully safely) cast it to what xarray is expecting, but this might let errors through.
-
-        ds = xr.open_dataset(
-            cast(XArrayOpenT, fpath),
-            drop_variables=drop_variables,
-            group=group,
-            decode_times=decode_times,
-        )
-
-        if indexes is None:
-            warnings.warn(
-                "Specifying `indexes=None` will create in-memory pandas indexes for each 1D coordinate, but concatenation of ManifestArrays backed by pandas indexes is not yet supported (see issue #18)."
-                "You almost certainly want to pass `indexes={}` to `open_virtual_dataset` instead."
-            )
-
-            # add default indexes by reading data from file
-            indexes = {name: index for name, index in ds.xindexes.items()}
-        elif indexes != {}:
-            # TODO allow manual specification of index objects
-            raise NotImplementedError()
-        else:
-            indexes = dict(**indexes)  # for type hinting: to allow mutation
-
-        # TODO we should drop these earlier by using drop_variables
-        loadable_vars = {
-            str(name): var
-            for name, var in ds.variables.items()
-            if name in loadable_variables
-        }
-
-        # if we only read the indexes we can just close the file right away as nothing is lazy
-        if loadable_vars == {}:
-            ds.close()
-    else:
-        loadable_vars = {}
-        indexes = {}
-
-    return loadable_vars, indexes
-
-
-def construct_virtual_dataset(
-    virtual_vars,
-    loadable_vars,
-    indexes,
-    coord_names,
-    attrs,
-) -> Dataset:
-    """Construct a virtual Datset from consistuent parts."""
-
-    vars = {**virtual_vars, **loadable_vars}
-
-    data_vars, coords = separate_coords(vars, indexes, coord_names)
-
-    vds = xr.Dataset(
-        data_vars,
-        coords=coords,
-        # indexes={},  # TODO should be added in a later version of xarray
-        attrs=attrs,
-    )
-
-    # TODO we should probably also use vds.set_close() to tell xarray how to close the file we opened
-
-    return vds
-
-
-def separate_coords(
-    vars: Mapping[str, xr.Variable],
-    indexes: MutableMapping[str, Index],
-    coord_names: Iterable[str] | None = None,
-) -> tuple[dict[str, xr.Variable], xr.Coordinates]:
-    """
-    Try to generate a set of coordinates that won't cause xarray to automatically build a pandas.Index for the 1D coordinates.
-
-    Currently requires this function as a workaround unless xarray PR #8124 is merged.
-
-    Will also preserve any loaded variables and indexes it is passed.
-    """
-
-    if coord_names is None:
-        coord_names = []
-
-    # split data and coordinate variables (promote dimension coordinates)
-    data_vars = {}
-    coord_vars: dict[
-        str, tuple[Hashable, Any, dict[Any, Any], dict[Any, Any]] | xr.Variable
-    ] = {}
-    for name, var in vars.items():
-        if name in coord_names or var.dims == (name,):
-            # use workaround to avoid creating IndexVariables described here https://github.com/pydata/xarray/pull/8107#discussion_r1311214263
-            if len(var.dims) == 1:
-                dim1d, *_ = var.dims
-                coord_vars[name] = (dim1d, var.data, var.attrs, var.encoding)
-
-                if isinstance(var, IndexVariable):
-                    # unless variable actually already is a loaded IndexVariable,
-                    # in which case we need to keep it and add the corresponding indexes explicitly
-                    coord_vars[str(name)] = var
-                    # TODO this seems suspect - will it handle datetimes?
-                    indexes[name] = PandasIndex(var, dim1d)
-            else:
-                coord_vars[name] = var
-        else:
-            data_vars[name] = var
-
-    coords = xr.Coordinates(coord_vars, indexes=indexes)
-
-    return data_vars, coords
