+++ virtualizarr/readers/dmrpp.py
@@ -9,33 +9,35 @@ import numpy as np
 import xarray as xr
 from xarray.core.indexes import Index
 
+from virtualizarr.backends.common import VirtualBackend
 from virtualizarr.manifests import ChunkManifest, ManifestArray
 from virtualizarr.types import ChunkKey
 from virtualizarr.utils import _FsspecFSFromFilepath
 from virtualizarr.zarr import ZArray
 
 
-def open_virtual_dataset(
-    filepath,
-    drop_variables: Iterable[str] | None = None,
-    loadable_variables: Iterable[str] | None = None,
-    indexes: Mapping[str, Index] | None = None,
-    reader_options: Optional[dict] = None,
-) -> xr.Dataset:
-    # TODO more sanitization of input types?
+class DMRPPVirtualBackend(VirtualBackend):
+    def open_virtual_dataset(
+        filepath,
+        drop_variables: Iterable[str] | None = None,
+        loadable_variables: Iterable[str] | None = None,
+        indexes: Mapping[str, Index] | None = None,
+        reader_options: Optional[dict] = None,
+    ) -> xr.Dataset:
+        # TODO more sanitization of input types?
 
-    if loadable_variables != [] or indexes is None:
-        raise NotImplementedError(
-            "Specifying `loadable_variables` or auto-creating indexes with `indexes=None` is not supported for dmrpp files."
-        )
+        if loadable_variables != [] or indexes is None:
+            raise NotImplementedError(
+                "Specifying `loadable_variables` or auto-creating indexes with `indexes=None` is not supported for dmrpp files."
+            )
 
-    fpath = _FsspecFSFromFilepath(
-        filepath=filepath, reader_options=reader_options
-    ).open_file()
-    parser = DMRParser(fpath.read(), data_filepath=filepath.strip(".dmrpp"))
-    vds = parser.parse_dataset()
-    vds.drop_vars(drop_variables)
-    return vds
+        fpath = _FsspecFSFromFilepath(
+            filepath=filepath, reader_options=reader_options
+        ).open_file()
+        parser = DMRParser(fpath.read(), data_filepath=filepath.strip(".dmrpp"))
+        vds = parser.parse_dataset()
+        vds.drop_vars(drop_variables)
+        return vds
 
 
 class DMRParser:
