+++ virtualizarr/tests/test_writers/test_zarr.py
@@ -6,7 +6,7 @@ import xarray.testing as xrt
 from xarray import Dataset
 
 from virtualizarr import ManifestArray, open_virtual_dataset
-from virtualizarr.backend import FileType
+from virtualizarr.backends.api import FileType
 from virtualizarr.manifests.manifest import ChunkManifest
 from virtualizarr.readers.zarr_v3 import metadata_from_zarr_json
 from virtualizarr.writers.zarr import dataset_to_zarr
